feat: migrate offline network popups to a global layout component
